# [DataTable] Optimize [NameNode/ConstNode] [=|!=|>|<|>=|<=] [NameNode/ConstNode] filtering