--- a/src/libraries/System.Data.Common/src/System/Data/Filter/BinaryNode.cs
+++ b/src/libraries/System.Data.Common/src/System/Data/Filter/BinaryNode.cs
@@ -319,7 +319,7 @@ namespace System.Data
                             Operators.LessThen => compareResult.Value < 0,
                             Operators.GreaterOrEqual => compareResult.Value >= 0,
                             Operators.LessOrEqual => compareResult.Value <= 0,
-                            Operators.NotEqual => (object)(compareResult.Value != 0),
+                            Operators.NotEqual => compareResult.Value != 0,
                             _ => throw ExprException.UnsupportedOperator(op),
                         };
                     }